+++ tests/test-suites/molecules/Container.tsx
@@ -5,7 +5,11 @@ import { Container } from "../../../src";
 const containerTest = () => {
   describe("Container", () => {
     test("Rendering test", () => {
-      let wrapper = mount(<Container>Test Container</Container>);
+      let wrapper = mount(
+        <Container animated animation="fade-in">
+          Test Container
+        </Container>
+      );
       expect(wrapper);
     });
   });